Banking Application

+++ src/main/java/net/javaguides/banking_app/Dto/TransactionDto.java
@@ -1,7 +1,8 @@
 package net.javaguides.banking_app.Dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
-public record TransactionDto(Long id, Long accountId, double amount, String transactionType,
+public record TransactionDto(Long id, Long accountId, BigDecimal amount, String transactionType,
                              LocalDateTime timestamp) {
 }
